move OrchestrationProperties to sharding-orchestration-core-facade module
