+++ sharding-orchestration/sharding-orchestration-core/sharding-orchestration-core-facade/src/main/java/org/apache/shardingsphere/orchestration/core/facade/properties/OrchestrationProperties.java
@@ -15,32 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.underlying.common.properties.orchestration;
+package org.apache.shardingsphere.orchestration.core.facade.properties;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.apache.shardingsphere.underlying.common.properties.common.TypedPropertyKey;
+import org.apache.shardingsphere.underlying.common.properties.common.TypedProperties;
+
+import java.util.Properties;
 
 /**
- * Typed property key of orchestration.
+ * Typed properties of orchestration.
  */
-@RequiredArgsConstructor
-@Getter
-public enum OrchestrationPropertyKey implements TypedPropertyKey {
+public final class OrchestrationProperties extends TypedProperties<OrchestrationPropertyKey> {
     
-    /**
-     * Enable or disable to overwrite orchestration config center data.
-     *
-     * <p>
-     * If this properties is true, yaml file config properties will overwrite orchestration config center data.
-     * Default: false
-     * </p>
-     */
-    OVERWRITE("overwrite", String.valueOf(Boolean.FALSE), boolean.class);
-    
-    private final String key;
-    
-    private final String defaultValue;
-    
-    private final Class<?> type;
+    public OrchestrationProperties(final Properties props) {
+        super(OrchestrationPropertyKey.class, props);
+    }
 }
